Working on a proposal system to submit a withdraw limit to other owners
This feature allows an owner (#1) to submit withdraw limits for the other owners (#2, #3) or himself.
For this, we need to track several requests and "actions" within the wallet:

 .A structure to keep track of an account settings;
.Mapping of who signed each settings proposal.
.Mapping similar to txMap, to track all the proposals submitted to other owners.
.Another mapping to connect an Account to an Address, after reaching the requested signatures.

Made the submitAccountLimit() that gives an owner the power to submit/propose a withdraw limit for the address given as parameter. Will store the proposer,
the last withdraw timestamp of the account, how many approvals were made, the new withdraw limit and to, the account passed in the function.

Started the signAccountLImit() function that will sign a proposal for the accounts.
Working on the withdraw function.