Building a withdraw limit proposal system. Each owner can submit a withdraw limit for other accounts, to be signed by the owners.  (#4)
In this branch I've built a structure to keep track of an account
withdraw timestamp and limit.
The `submitAccountLimit` function lets each owner submit proposals for
withdraw limits in others accounts, the proposals will be executed if
the required signatures are met.
Like in a transaction, I used a mapping to store all proposals indexes.
The `signSettings` function, allows owners to sign on a specific
proposal. This will call an internal function that executes all the
logic of signing a proposal.
Made a `withdraw` function to be called by an owner. This owner can only
withdraw up to the withdraw limit.